Update GameCursor.cs
apply cursor size setting

the 4 is determined from the following: medium is the default, and 64x64 is assumed to be medium for 1080p. 1080/64 is about 16. each size setting is double/half the adjacent setting, so the smallest setting would be 4.
this may be bumped up in the future as things are tested